requirements.txt
Django==2.2.6
gunicorn==20.0.0
mock==3.0.5
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
